refactor: move account service to its dir
